+++ kamehouse-shell/bin/common/functions/kamehouse/deployment-functions.sh
@@ -126,7 +126,7 @@ deployKameHouseMobile() {
 
 uploadKameHouseMobileApkToHttpdServer() {
   log.info "Deploying ${COL_PURPLE}kamehouse-mobile${COL_DEFAULT_LOG} app to downloads server"
-  log.info "Set KAMEHOUSE_MOBILE_APP_SERVER, KAMEHOUSE_MOBILE_APP_PATH and KAMEHOUSE_MOBILE_APP_USER in kamehouse.cfg"
+  log.info "Set ${COL_RED}KAMEHOUSE_MOBILE_APP_SERVER, KAMEHOUSE_MOBILE_APP_PATH and KAMEHOUSE_MOBILE_APP_USER${COL_DEFAULT_LOG} in ${HOME}/.kamehouse/kamehouse.cfg"
   log.debug "scp -v ${KAMEHOUSE_ANDROID_APK_PATH} ${KAMEHOUSE_MOBILE_APP_USER}@${KAMEHOUSE_MOBILE_APP_SERVER}:${KAMEHOUSE_MOBILE_APP_PATH}/kamehouse.apk"
   scp -v ${KAMEHOUSE_ANDROID_APK_PATH} ${KAMEHOUSE_MOBILE_APP_USER}@${KAMEHOUSE_MOBILE_APP_SERVER}:${KAMEHOUSE_MOBILE_APP_PATH}/kamehouse.apk
   checkCommandStatus "$?" "An error occurred deploying kamehouse-mobile through ssh"