+++ kamehouse-shell/bin/kamehouse/kamehouse-mobile-update-apk-status-html.sh
@@ -12,8 +12,8 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
   exit 99
 fi
+source ${HOME}/.kamehouse/kamehouse.cfg
 
-KAMEHOUSE_MOBILE_APP_PATH="/var/www/kamehouse-webserver/kame-house-mobile"
 KAMEHOUSE_APK_HTML_TEMPLATE=${HOME}/programs/kamehouse-shell/conf/kamehouse-apk-status-template.html
 KAMEHOUSE_APK_HTML=kamehouse-apk-status.html
 BUILD_VERSION=""