+++ kamehouse-shell/bin/kamehouse/kamehouse-mobile-upload-apk-to-device.sh
@@ -17,13 +17,11 @@ source ${HOME}/.kamehouse/.shell/shell.pwd
 source ${HOME}/.kamehouse/kamehouse.cfg
 
 SKIP_BUILD_MOBILE=false
-SD_CARD_APK_PATH=/0/Download
-SFTP_USER=android
 USE_CURRENT_DIR=true
 
 mainProcess() {
   setKameHouseRootProjectDir
-  log.info "Set ANDROID_IP and ANDROID_PORT in kamehouse.cfg"
+  log.info "Set ${COL_RED}ANDROID_IP, ANDROID_PORT, ANDROID_SFTP_USER, ANDROID_APK_DEST_PATH${COL_DEFAULT_LOG} in ${HOME}/.kamehouse/kamehouse.cfg"
   if ${SKIP_BUILD_MOBILE}; then
     log.info "Running with -s. Skipping build kamehouse-mobile"
   else
@@ -40,13 +38,13 @@ uploadApkToDeviceSftp() {
   log.info "${COL_PURPLE}Uploading${COL_DEFAULT_LOG} kamehouse-mobile apk ${COL_PURPLE}to android device${COL_DEFAULT_LOG} through sftp"
   log.info "Check pass in sftp mobile app config and store it in ${HOME}/.kamehouse/.shell/shell.pwd as ANDROID_SFTP_PASS=password ${COL_PURPLE}to execute without password prompt"
   if ${IS_LINUX_HOST}; then 
-    log.debug "sftp -v -P ${ANDROID_PORT} ${SFTP_USER}@${ANDROID_IP} <<< \"put ${KAMEHOUSE_ANDROID_APK_PATH} ${SD_CARD_APK_PATH}/kamehouse.apk\" "
-    sftp -v -P ${ANDROID_PORT} ${SFTP_USER}@${ANDROID_IP} <<< "put ${KAMEHOUSE_ANDROID_APK_PATH} ${SD_CARD_APK_PATH}/kamehouse.apk" 
+    log.debug "sftp -v -P ${ANDROID_PORT} ${ANDROID_SFTP_USER}@${ANDROID_IP} <<< \"put ${KAMEHOUSE_ANDROID_APK_PATH} ${ANDROID_APK_DEST_PATH}/kamehouse.apk\" "
+    sftp -v -P ${ANDROID_PORT} ${ANDROID_SFTP_USER}@${ANDROID_IP} <<< "put ${KAMEHOUSE_ANDROID_APK_PATH} ${ANDROID_APK_DEST_PATH}/kamehouse.apk" 
   else
     log.warn "Putty pscp needs to be installed. if not switch to standard scp. Run with log=debug to see scp command"
-    log.debug "sftp -v -P ${ANDROID_PORT} ${SFTP_USER}@${ANDROID_IP} <<< \"put ${KAMEHOUSE_ANDROID_APK_PATH} ${SD_CARD_APK_PATH}/kamehouse.apk\" "
-    log.debug "pscp -pw [pass] -v -P ${ANDROID_PORT} ${KAMEHOUSE_ANDROID_APK_PATH} ${SFTP_USER}@${ANDROID_IP}:${SD_CARD_APK_PATH}/kamehouse.apk"
-    pscp -pw ${ANDROID_SFTP_PASS} -v -P ${ANDROID_PORT} ${KAMEHOUSE_ANDROID_APK_PATH} ${SFTP_USER}@${ANDROID_IP}:${SD_CARD_APK_PATH}/kamehouse.apk
+    log.debug "sftp -v -P ${ANDROID_PORT} ${ANDROID_SFTP_USER}@${ANDROID_IP} <<< \"put ${KAMEHOUSE_ANDROID_APK_PATH} ${ANDROID_APK_DEST_PATH}/kamehouse.apk\" "
+    log.debug "pscp -pw [pass] -v -P ${ANDROID_PORT} ${KAMEHOUSE_ANDROID_APK_PATH} ${ANDROID_SFTP_USER}@${ANDROID_IP}:${ANDROID_APK_DEST_PATH}/kamehouse.apk"
+    pscp -pw ${ANDROID_SFTP_PASS} -v -P ${ANDROID_PORT} ${KAMEHOUSE_ANDROID_APK_PATH} ${ANDROID_SFTP_USER}@${ANDROID_IP}:${ANDROID_APK_DEST_PATH}/kamehouse.apk
   fi
 }