+++ changelog.md
@@ -12,8 +12,9 @@
 - Renamed kamehouse config files
 - Made the vlc video playlists path configurable
 - Removed dependency on media server hostname for accessing playlists
-- Moved hardcoded configured kamehouse servers to kamehouse.cfg
-- Moved docker host ip to be set in kamehouse.cfg
+- Moved hardcoded configured kamehouse servers to ${HOME}/.kamehouse/kamehouse.cfg
+- Moved docker host ip to be set in ${HOME}/.kamehouse/kamehouse.cfg
+- Moved mobile app servers list to ${HOME}/.kamehouse/kamehouse.cfg
 - Fixed bugs
 
 #### v9.02